+++ src/test/java/org/folio/search/service/setter/item/ItemTypedCallNumberProcessorTest.java
@@ -4,12 +4,7 @@ import static org.apache.commons.collections4.SetUtils.emptySet;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.search.client.InventoryReferenceDataClient.ReferenceDataType.CALL_NUMBER_TYPES;
 import static org.folio.search.model.client.CqlQueryParam.SOURCE;
-import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.DEWEY_ID;
-import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.LC_ID;
 import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.LOCAL_CALL_NUMBER_TYPES_SOURCES;
-import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.NLM_ID;
-import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.OTHER_SCHEME_ID;
-import static org.folio.search.service.setter.item.ItemTypedCallNumberProcessor.SUDOC_ID;
 import static org.mockito.Mockito.when;
 
 import java.util.List;
@@ -19,6 +14,7 @@ import java.util.stream.Stream;
 import org.folio.search.domain.dto.Instance;
 import org.folio.search.domain.dto.Item;
 import org.folio.search.integration.ReferenceDataService;
+import org.folio.search.model.types.CallNumberType;
 import org.folio.spring.test.type.UnitTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -37,26 +33,6 @@ class ItemTypedCallNumberProcessorTest {
 
   private @InjectMocks ItemTypedCallNumberProcessor processor;
 
-  private static Instance instance(Item... items) {
-    return new Instance().items(List.of(items));
-  }
-
-  private static Item item(String effectiveShelvingOrder, String callNumberTypeId) {
-    return new Item().effectiveShelvingOrder(effectiveShelvingOrder).itemLevelCallNumberTypeId(callNumberTypeId);
-  }
-
-  public static Stream<Arguments> emptyCallNumberAfterProcessingSource() {
-    return Stream.of(
-      Arguments.arguments(null, null),
-      Arguments.arguments("", ""),
-      Arguments.arguments("", LC_ID),
-      Arguments.arguments("AAA", ""),
-      Arguments.arguments(null, LC_ID),
-      Arguments.arguments("AAA", null),
-      Arguments.arguments("()[]", LC_ID)
-    );
-  }
-
   @Test
   void getFieldValue_multipleValue_positive() {
     var localCnId = UUID.randomUUID().toString();
@@ -65,19 +41,19 @@ class ItemTypedCallNumberProcessorTest {
       .thenReturn(Set.of(localCnId));
     var eventBody = instance(
       item("HD 11", localCnId),
-      item("HD 11", LC_ID),
-      item("HD 11", DEWEY_ID),
-      item("HD 11", NLM_ID),
-      item("HD 11", SUDOC_ID),
-      item("HD 11", OTHER_SCHEME_ID)
+      item("HD 11", CallNumberType.LC.getId()),
+      item("HD 11", CallNumberType.DEWEY.getId()),
+      item("HD 11", CallNumberType.NLM.getId()),
+      item("HD 11", CallNumberType.SUDOC.getId()),
+      item("HD 11", CallNumberType.OTHER.getId())
     );
     var actual = processor.getFieldValue(eventBody);
-    assertThat(actual).containsExactly(84787310808212480L,
-      229342416757269504L,
+    assertThat(actual).containsExactly(229342416757269504L,
       373897522706326528L,
       518452628655383552L,
       663007734604440576L,
-      807562840553497600L);
+      807562840553497600L,
+      952117946502554624L);
   }
 
   @ParameterizedTest
@@ -95,5 +71,24 @@ class ItemTypedCallNumberProcessorTest {
     var eventBody = instance(item("AA 123", UUID.randomUUID().toString()));
     var actual = processor.getFieldValue(eventBody);
     assertThat(actual).isEmpty();
+  }
+
+  public static Stream<Arguments> emptyCallNumberAfterProcessingSource() {
+    return Stream.of(
+      Arguments.arguments(null, null),
+      Arguments.arguments("", ""),
+      Arguments.arguments("", CallNumberType.LC.getId()),
+      Arguments.arguments("AAA", ""),
+      Arguments.arguments(null, CallNumberType.LC.getId()),
+      Arguments.arguments("AAA", null)
+    );
+  }
+
+  private static Instance instance(Item... items) {
+    return new Instance().items(List.of(items));
+  }
+
+  private static Item item(String effectiveShelvingOrder, String callNumberTypeId) {
+    return new Item().effectiveShelvingOrder(effectiveShelvingOrder).itemLevelCallNumberTypeId(callNumberTypeId);
   }
 }